Fix locale translations for Nostr posts
